Refactor EmailService: consolidate email functionality, remove redundant class, and update imports
