--- govconnect-server/src/main/java/com/novaix/govconnect_server/service/email/EmailService.java
@@ -1,54 +1,0 @@
-package com.novaix.govconnect_server.service.email;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.stereotype.Service;
-import org.thymeleaf.TemplateEngine;
-import org.thymeleaf.context.Context;
-
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class EmailService {
-
-    private final JavaMailSender mailSender;
-    private final TemplateEngine templateEngine;
-
-    @Value("${spring.mail.username}")
-    private String fromEmail;
-
-    public void sendOtpEmail(String toEmail, String otp, LocalDateTime expiryTime) {
-        try {
-            MimeMessage message = mailSender.createMimeMessage();
-            MimeMessageHelper helper = new MimeMessageHelper(message, true);
-
-            helper.setFrom(fromEmail);
-            helper.setTo(toEmail);
-            helper.setSubject("Password Reset OTP - GovConnect");
-
-            Context context = new Context();
-            context.setVariable("otp", otp);
-            context.setVariable("expiryTime", expiryTime.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss")));
-            context.setVariable("validMinutes", "5");
-
-            String htmlContent = templateEngine.process("forgot-password-otp-new", context);
-            helper.setText(htmlContent, true);
-
-            mailSender.send(message);
-            log.info("OTP email sent successfully to: {}", toEmail);
-
-        } catch (MessagingException e) {
-            log.error("Failed to send OTP email to: {}", toEmail, e);
-            throw new RuntimeException("Failed to send email", e);
-        }
-    }
-}
