+++ govconnect-server/src/test/java/com/novaix/govconnect_server/service/password/PasswordResetServiceTest.java
@@ -34,7 +34,7 @@ import com.novaix.govconnect_server.request.ForgotPasswordRequest;
 import com.novaix.govconnect_server.request.ResetPasswordRequest;
 import com.novaix.govconnect_server.response.ForgotPasswordResponse;
 import com.novaix.govconnect_server.response.ResetPasswordResponse;
-import com.novaix.govconnect_server.service.email.EmailService;
+import com.novaix.govconnect_server.service.common.EmailService;
 
 @ExtendWith(MockitoExtension.class)
 class PasswordResetServiceTest {
